Code cleanup + CodeFactor Suggestions
Removed redundant constructor from ActivitiesController
Removed blank line from Startup
Seperated IsHostRequirement and IsHostRequirementHandler into different files
Removed blank line from DataContext
