+++ Persistance/DataContext.cs
@@ -8,7 +8,6 @@ namespace Persistance
     {
         public DataContext(DbContextOptions options) : base(options)
         {
-
         }
 
         public DbSet<Activity> Activities { get; set; }
